Small refactor to, some prep to support multiple languages in the future
